requirements.txt
anyio==4.8.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.12.3
bertopic==0.15.0
black==25.1.0
bleach==6.2.0
bs4==0.0.2
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docker==4.4.4
executing==2.2.0
fake-useragent==2.0.3
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.17.0
flake8==7.1.1
fonttools==4.55.8
fqdn==1.5.1
frozendict==2.4.6
fsspec==2025.2.0
h11==0.14.0
hdbscan==0.8.40
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
identify==2.6.6
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.0.3
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
llvmlite==0.44.0
loguru==0.5.3
lxml==5.3.0
MarkupSafe==2.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.1.1
mpmath==1.3.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.24.2
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nitter-scraper==0.5.0
nltk==3.9.1
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.24.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pathspec==0.12.1
peewee==3.17.8
pendulum==2.1.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
praw==7.8.1
prawcore==2.4.0
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pycparser==2.22
pydantic==1.10.21
pyee==11.1.1
pyflakes==3.2.0
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
pyzmq==26.2.0
ratelimit==2.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-html==0.10.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.28.1
Send2Trash==1.8.3
sentence-transformers==3.4.1
six==1.17.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.2
trio==0.28.0
trio-websocket==0.11.1
triton==3.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
undetected-chromedriver==3.5.5
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
vaderSentiment==3.3.2
virtualenv==20.29.1
w3lib==2.3.1
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
widgetsnbextension==4.0.13
wsproto==1.2.0
zipp==3.21.0
